refactor(kubernetes): Remove KubernetesCachingAgent interface
Push everything down to KubernetesV2CachingAgent, which is the only
implmentation now that the V1 provider is deleted.
